Cleanup

+++ src/algebra/Matrix.h
@@ -23,8 +23,8 @@ template <typename T>
 class Matrix {
 
   /// Index from a 1D array to a position in a 2D array (matrix)
-  long global2Local(const long& row, const long& col);
-  std::tuple<long, long> local2Global(const long& k);
+  int global2Local(const int& row, const int& col);
+  std::tuple<int, int> local2Global(const int& k);
 
  /** Boolean variable which tells us if the underlying matrix is parallel
  * will be defaulted to false if no value is provided in constructor
@@ -72,7 +72,7 @@ class Matrix {
   /** Find the global indices of the matrix elements that are stored locally
    * by the current MPI process.
    */
-  std::vector<std::tuple<long, long>> getAllLocalStates();
+  std::vector<std::tuple<int, int>> getAllLocalStates();
 
   /** Returns true if the global indices (row,col) identify a matrix element
    * stored by the MPI process.
@@ -81,17 +81,17 @@ class Matrix {
 
   /** Find global number of rows
    */
-  long rows() const;
+  int rows() const;
   /** Return local number of rows */
-  long localRows() const;
+  int localRows() const;
   /** Find global number of columns
    */
-  long cols() const;
+  int cols() const;
   /** Return local number of rows */
-  long localCols() const;
+  int localCols() const;
   /** Find global number of matrix elements
    */
-  long size() const;
+  int size() const;
 
   /** Get and set operator
    */
@@ -247,26 +247,26 @@ Matrix<T>::~Matrix() {
 
 /* ------------- Very basic operations -------------- */
 template <typename T>
-long Matrix<T>::rows() const {
+int Matrix<T>::rows() const {
   if(isDistributed) return pmat->rows();
   else{ return mat->rows(); }
 }
 template <typename T>
-long Matrix<T>::localRows() const {
+int Matrix<T>::localRows() const {
   if(isDistributed) return pmat->localRows();
   else{ return mat->rows(); }
 }
-template <typename T> long Matrix<T>::cols() const {
+template <typename T> int Matrix<T>::cols() const {
   if(isDistributed) return pmat->cols();
   else{ return mat->cols(); }
 }
 template <typename T>
-long Matrix<T>::localCols() const {
+int Matrix<T>::localCols() const {
   if(isDistributed) return pmat->localCols();
   else{ return mat->cols(); }
 }
 template <typename T>
-long Matrix<T>::size() const {
+int Matrix<T>::size() const {
   if(isDistributed) return pmat->size();
   else{ return mat->size(); }
 }
@@ -291,20 +291,20 @@ bool Matrix<T>::indicesAreLocal(const int& row, const int& col) {
 }
 
 template <typename T>
-std::tuple<long, long> Matrix<T>::local2Global(const long& k) {
+std::tuple<int, int> Matrix<T>::local2Global(const int& k) {
   if(isDistributed) return pmat->local2Global(k);
   else{ return mat->local2Global(k); }
 }
 
 // Indexing to set up the matrix in col major format
 template <typename T>
-long Matrix<T>::global2Local(const long& row, const long& col) {
+int Matrix<T>::global2Local(const int& row, const int& col) {
   if(isDistributed) return pmat->global2Local(row,col);
   else{ return mat->global2Local(row,col); }
 }
 
 template <typename T>
-std::vector<std::tuple<long, long>> Matrix<T>::getAllLocalStates() {
+std::vector<std::tuple<int, int>> Matrix<T>::getAllLocalStates() {
   if(isDistributed) return pmat->getAllLocalStates();
   else{ return mat->getAllLocalStates(); }
 }
