+++ src/bands/window.h
@@ -53,21 +53,21 @@ class Window {
   // Constants that identify the kind of filter to be used
   /** nothing=0 identifies the do-nothing window
    */
-  static const long nothing = 0;
+  static const int nothing = 0;
 
   /** population labels the window type looking for partially occupied states
    */
-  static const long population = 1;
+  static const int population = 1;
 
   /** energy labels the window type looking for states within two energy values
    */
-  static const long energy = 2;
+  static const int energy = 2;
 
   /** Returns the kind of energy filter used.
    * @return method: an integer equal to either Window::nothing,
    * Window::population, or Window::energy.
    */
-  long getMethodUsed();
+  int getMethodUsed();
  private:
   /** particle stores whether we are working with electrons or phonons
    */
@@ -85,7 +85,7 @@ class Window {
 
   /** temp variable to facilitate code writing
    */
-  long numBands;
+  int numBands;
 
   // internal method to apply the window on population
   std::tuple<std::vector<double>, std::vector<int>> internalPopWindow(
